requirements.txt
unstructured
requests
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy
python-dotenv
bs4
sentence-transformers
fastapi==0.95.1
uvicorn==0.21.1
